+++ src/app/(dashboard)/account/schemas/password-form.schema.ts
@@ -2,10 +2,11 @@ import { z } from 'zod';
 
 export const AccountPasswordFormSchema = z
   .object({
-    password: z.string().min(8, 'Password needs to be longer than 8 characters'),
+    currentPassword: z.string().min(8, 'Please enter your current password'),
+    newPassword: z.string().min(8, 'Your new password needs to be longer than 8 characters'),
     confirmPassword: z.string(),
   })
-  .refine((data) => data.password === data.confirmPassword, {
+  .refine((data) => data.newPassword === data.confirmPassword, {
     message: 'Passwords do not match',
     path: ['confirmPassword'],
   });